Original LS1021a-TWR device tree taken from linux
These are the original device tree source files as provided by
NXP together with the LS1021a-TWR. The files are part of the
git repository provided on an USB stick and are taken from sha id
aaa395f27e077f51c75c6aefc7bbec44732e9742. A more recent release
contains a slightly different tree, but we stick with this version
for the time being since it matches the currently used sources.

The files are added as a reference for future changes

Change-Id: Ibf88f416691b3e160c42c1cd1295e3385353166d