Find the proper 'compatible' device
When looking for a device based on the compatible entry in a
Dt_node, we might pick the wrong one out of a set of devices
with the same compatible list (e.g. in Io_proxy's
Factory create function). This leads to the situation that
GPA are not equal to HPA, i.e. one works on the wrong
device instance.
To fix that we also compare for resource entries to pick the
right device.

Change-Id: I8a2f77db3e6c98478b046a53dec1ad5bc21c7e0d