Keep track of unaligned region start
Ds_handler manages page aligned objects. It calculates the page to be
mapped after a page fault as follows:

	addr = local_start + (pfa - region_start)

(pfa - region_start is passed as parameter offset)
To get the correct page we have to consider the offset within the
page which is part of region_start. Therefore we add the offset
with a page to local_start after we mapped the data space.

TODO: We currently use the unaligned addresses to attach to the data
      space. attach() passes the unaligned address to IO to request
      the region and IO uses this address to look for the requested
      region. If we align start and size before invoking attach IO
      does not find the region and the attach fails. This needs to
      be fixe together with IO.

      This change does not solve the problem for the MAP_OTHER case.
      Here we would have to fix the start of the region or reject a
      non aligned mapping.

Change-Id: I2f52933fa116a4471977c90281fc41ef2784bd2b