mips: remove GIC core-local area from mmap area
The core-local registers in the Mips GIC provide a per CPU view.
That means they are different for each VCPU and therefore cannot
be mapped into the common task.

This change allows the mmapped mmio device to have a smaller size
then what is covered by the actual device registers. GIC uses that
to only map the shared registers into the guest. Core-local registers
are handled as usual.

Also slightly simplifies the constructor of the GIC. The size parameter
is no longer necessary. Directly hand in the core IC instead which is
never set anywhere else after construction.

Change-Id: I71283814ecf84b22dd87791287e2489684d2b921