mips: fix access size computation and ignore 8/16 bit access to GIC
Adapt the convention of ARM's sas bits regarding access size:
the values 0-3 are used meaning 8/16/32/64 bit access. That
fixes an off-by-1 bug in the Mips instruction decoding of
access width.

Ignore 8 and 16 bit access to mips GIC. They were in theory
possible before when they were aligned to a 32bit boundary.
In practice the code was buggy with respect to sign extension.
32bit access on a 64bit architecture is still possible.

Change-Id: Ic9e230e5acbf86168354a400cc1b5da8da03f871