mips: save/restore FPU register in exit handler
The mips64 compiler agressively uses FPU registers as general
purpose registers. In particular many of the l4re support libraries
are affected. As long as we don't have soft-float versions of these
libraries, the FPU registers need to be saved on each entry.

The entry function has been moved into a separate file which is
compiled with soft-float. That allows to use standard C++ code
to implement FPU save/restore.

Change-Id: If3b3cbe41e053386f008042ebfe43eb3e0665374