mips: implement mtc0 counter and a more precise wakeup from wait
A counter change causes guest_timer_offset to be set relative to the
hardware clock visible to the vmm.

The wait timeout is now computed relative to the kip clock time
when the hardware clock was read. The kip clock is read before and
after reading the hardware clock to make sure the thread hasn't
been interrupted between reading the two clocks. Round up timeout
by KIP clock granualarity as we don't have any information how
hardware clock and KIP clock are synchronised and we want to make
sure the timer has hit when the receive timeout hits.

Change-Id: Iacc038e632cea0f90ec9383d6837cbf22a8e2b3f