Add/Update psci node on arm
The psci node describes the interface to a secure monitor or hypervisor.
The uvmm protocol and service call might be different from the entry
in the device tree (e.g. if the device tree is provided by Uboot, which
adds its own protocol entries). So we have to add/update the psci
node before passing the device tree to the VM.

The update is done by deleting/re-adding the node (instead of iterating
over all psci node properties to remove them)

Change-Id: I3d2729f026d053fb7f57f7a259a83998e21c9024