Check local and shared irqs in schedule_irqs
Due to the limited amount of list registers it might happen, that
we can not forward an irq to the guest. Such an irq will stay
in a pending state and needs to be picked up when checking for
pending interrupts. schedule_irqs() checked only spi irqs and
missed cpu local irqs which led to lost timer irqs (which are
cpu local).

Here we add a check for local irqs.

Change-Id: I98e72b99bc00599dc6a6d985199ff5a58983453f