Re-use allocated list register
List registers might be in one of 4 states:
* invalid
* pending
* active
* active and pending

If an interrupt comes in it might be, that the irq is currently
handled by the guest and therefore is in an active state. In this
situation we have to change the state of the list register from
"active" to "active and pending".

This allows for an immediate delivery of the pending IRQ after
an EOI. It also prevents an inconsistency in the IRQ state. If we
use a new list register for an active IRQ we get the EOI twice.
On the first EOI the IRQ is marked as not active on any CPU, on
the second EOI we realize, that we got an EOI for an IRQ which
is not active on any CPU and trigger an assertion.

To implement this we keep track of the used list register and check
it before allocating an empty one. If we find a used one we accept
the irq on this cpu and set the pending state bit.

This should fix the problems described by Frank in ticket #140.

Change-Id: I515592192f7e390cda543e4951609aafba4b1b4b