ARM: new Fiasco API running ARM VMM in PL0/EL0
On ARM, Fiasco used to switch the thread/vCPU with full virtualization
support to PL1/EL1 (SVC mode). With this change the VMM for ARM runs at
PL0 (USR mode) as on any other architecture.  This has very little
consequences for our uvmm, however, if a VMM wnat access to the full
state of a running VM it has to use a VM-clear system call and VM-read
VM-write APIs.

Change-Id: I3b074c5ac4cb8448181940cb21a9971f485f08a5