Calculate timeouts based on run-time information
Information about the current timer frequency is found in
* device tree
* timer frequency register
  * CNTFRQ read via mrc/mrs
  * CNTFRQ read using mmio

Here we read the rate from CNTFRQ (mrc/mrs only) and derive
values used to convert timer ticks to micro seconds.

If there is a clock frequency property we compare it with the currently
used one and update it if necessary.

Change-Id: I96544b440d56f88b583a34ceade819e865f99466