ARM: Fix Linux image detection
Detection for 64bit Linux kernels checked for the PE
header. This is not correct because a 32bit EFI image
has the same header. Instead we check for specific
Linux header. Also, a 64bit Linux without EFI support
does not have the header as well and would not have
been detected correctly. The Linux specific headers
are there both with and without included EFI support.

Change-Id: Ibf096d9185e1f175fe62caae58bbcd4b5b0a372e