collect all static parts and devices in new Vm class
The aim is that the Vm instance is the central point where any
device in the system can find references to other devices. We
already do that for devices in the device tree with device repo.
VM allows to find the statically created devices (vmm, cpus, vbus)
as well and simplifies dependency handling throughout the code.
Vmm::Guest on the other hand is being degraded to the class that does
the VM exit handling only.

Change-Id: I1c0df8b4b040f77a7415b4ffa3d9b71d1b0b4f0c