Consider memory alignment while calculating page sizes
We try to map superpage sized objects while resolving a page-fault.
To be able to do that we have to consider some restrictions required
by flexpages:

* regions in the address space of the virtual machine and the
  uvmm have to have an alignment which is compatible with the
  flexpage size
* the page must be a subset of the mapped region

Here we introduce get_page_shift() which looks at both regions
and calculates a pagesize fulfilling the requirements.
Currently we only return either L4_PAGESIZE or L4_SUPERPAGESIZE.
This might be replaced in the future with a function returning
the largest possible log2(pagesize) possible for the region.

Change-Id: Id245494872cf21fd825ff5e13d7c6a9a74bf1230