Write VQ config directly from the transport connector
The transport connector writes the virtqueue configuration for the
descriptor, available, and used ring into the device header config.
If the VQ is enabeld, it reads this config.
In the case where the driver configures all VQs and then enables
them all, the last configured VQ values written to the header config,
will be taken to configure all VQs.

To prevent this, the connector writes directly to the queue config
in the device header config page (see l4virtio_config_queues).
Furthermore, the VQ is set up using the corresponding values from this
table.

To manage access to the virtqueues Virtio::Dev requires derived
device classes to implement an accessor function for the device's
virtqueues.

Change-Id: I1db44bd811cae5bf6923abdb623f49d81661cc8f