PCI and Virtio-PCI transport
The PCI bus bridge is capable to serve a PCI bus with purely emulated
devices or a mix of IO provided and emulated PCI devices to the guest.

PCI configuration space access via IO ports is implemented in a
separate component component talking an internal protocol with the PCI
bus bridge, which should be able to serve as configuration space
access methods as well.

All virtio PCI devices are multi-function devices per default.
The Pci_bus_bridge is not.

Change-Id: Ie180368dbb8efbfbb3637f2d07ef2ea6356768de