virtio-proxy: only use device-tree-supplied config page size
The dataspace received from the registration function may
be larger than the config size configured in the device tree.
It is sufficient to only map the part needed there. Also
adds a check that the dataspace is large enough and uses
the info() function directly to make sure to compare against
the full unsigned tange.

Change-Id: I0f18440482726dd8936125106ade352e2cd80a06