Add additional CPUs
We introduced SMP for MIPS and changed ARM to boot CPU0 using the
new interfaces introduced to setup and boot bultiple CPUs. Here
we add the missing pieces to boot multiple CPUs on ARM.

We create one thread per virtual CPU. We prepare the register state
and activate it on the physical CPU it is supposed to run on. There
the thread finishes the initialization of the virtual CPU and enters
the virtual machine. CPU0 is powered up by default, the other CPUs
are started via PSCI.

With this change an SMP Linux starts the other CPU core that are
available but stops there since the other CPUs do not respond to IPIs,
which are added in a seperate commit.

TODO:
* add the complete PSCI interface

Change-Id: I8455a6b667dcf3c783daa8e9746c895b62f567dc