Move cpu specific setup to Cpu_dev constructor
During setup of Cpu_dev we extract information about the CPU from
the device tree. The information extracted depends on the
architecture:
* all:   cpu number (reg property)
* amd64: none
* mips:  compatibility attributes (compatible property)
* arm:   MPIDR value (reg property)

To support this architecture specific setup we pass the device tree
node to the constructor. set_proc_type() is not needed anymore and is
removed.

A default CPU is only created if there is no CPU created from device
tree nodes.

The CPU number contained in the reg property might be encoded in a
architecture specific way. Therefore we add an architecture specific
dtid_to_cpuid() function, which can be used to extract the CPU number
from the reg property value.

Change-Id: I1bfcd59464efd7ccba48281b8e03498d5813089a