Set VMPIDR for VCPU
Guests read topology/affinity information from MPIDR and get confused
if the values read differ from their expectation. Therefore we set the
VMPIDR to the values shown in the device three (if a valid cpu node
is present, otherwise we use the assigned logical cpu number).

We use the state set by fiasco and replace the affinity information.
We reset the UP/MT bit (multi threading) and keep the other bits as
they are (Multiprocessor extension bit and Res0 bits).

Change-Id: Ie2ed61b0665ad7c0b204b5d9c44bc5dc898e6826
Requires-Change-Id: I7bc34a4aa2667154cd206455120d44495c556eb9