Seperate hardware ids and logical CPU numbers
The device three for arm contains hardware ids for each CPU node. The
hardware id matches the contents of the MPIDR registers and contains
topology information.

Linux uses the hardware id to find device tree nodes for CPUs and to
identify CPUs in psci calls. Here we seperate the logical numbering
of CPUs and the management of hardware ids. We simply enumerate CPUs
as they are listed in the device tree and use this number to identify
CPUs internally (as index into the cpu array) and to enumerate the
GIC CPU interfaces.

The hardware id is kept in cpu_dev to
* setup the VMPIDR to allow linux to identify the CPU and
* to find the CPU mentioned in psci calls

We might replace the hardware id with completely made up ids at some
later point. We just have to make sure that the reg entry of the CPU
node and the VMPIDR contain the same value. Additionally we have to
consider that Linux derives scheduling information from the topology
information in the hw id.

Change-Id: Id9ed1448bb10f105d825c30ab57bff31cd4b51d5