amd64: Fix instruction decoding for MMIO
MAD did decode access to the lower byte of SP, BP, SI, and DI to
reference AH, BH, CH, DH. This is correct for 32bit access but not for
64bit access where a REX prefix is present in the instruction.
In this case, SPL, BPL, SIL, and DIL must be used.

vcpu_ptr.decode: The translation from MAD description to Mem_access
ignored the shift value in the MAD descriptor.

Change-Id: I221d8203fadcc02d6460b4bbe0696e59eca6a87f