mmio_proxy: Allow for smaller Dataspace
Allow that we do not know the exact size of a dataspace,
i.e. consider that we give a MMIO range in the DT
that is larger than the dataspace we back the
area with. For that, we take the smaller value
of both the MMIO area and the dataspace size.
The guest should not access beyond the size of the
dataspace despite the indication of the max size in
the DT.

Change-Id: Ib583b52f934536bad202cbeb29bc1a464f9f43c9