Request memory using RPC instead of touching it
To be able to map memory into the address space of the VM the memory has to
be present in address space of uvmm. The usual way to guarantee this was
to access the memory with read/write operations. But this has some side
effects, e.g. creating dirty cache lines which might be written back at an
arbitrary point in time and might overwrite data in the VM written directly
to memory.

Here we replace the touch operations by an explicit request to the region
manager. To ensure that we do not request memory that is already there we
* attach to data spaces with Eager_map
* map all memory to the address space of the VM before starting it
A memory related exit from the VM the means, that for some reasons the
memory is no present anymore and we have to request it again.

Changes:
* mmio_device.h
  * page_in() - manually request page fault resolution to establish a mapping
                for the specified address and mode
  * map_guest_range() - Map address range into guest
  * map_eager() - method invoked to map a region into the VM, default noop
  * map_mmio() - replace touch_ro with explicit request for memory
* ds_mmio_mapper.h
  * provide map_eager()
  * request memory before mapping it to the VM address space
  * always map memory RWX
    * since we map memory RWX after a write access there is no reason to
      map it read only after a read access
* ram_ds.h
  * remove touch_rw()
* virtio_dev.h, virtio_proxy.h
  * add Eager_map option to attach
* generic_guest.h
  * map_eager() - iterate over all regions in memory map and invoke
    map_eager() on each of them
* main.c
  * remove touch_rw()
  * invoke map_eager() on guest vmm

Fixes #708

Change-Id: I5590910e691a790de804c3fc91a99877ccb980a9