Filter out root IRQ parent
When we try to lookup the IRQ parent for the root IRQ controller
we might find the same IRQ controller again. If we return the node
itself the caller might end up in an endless loop while traversing
the IRQ controller hierarchy.

Here we add an additional check to return an invalid node if the
node is the root IRQ controller and remove the identity check in
create_irq_parent().

Change-Id: I4f60894626b258a9576c988efc157b5582533686