Introduce interrupt parent helper method
Here we introduce two methods to get or create the interrupt parent:
* Device_lookup::get_or_create_ic_dev(node, fatal)
  * returns the interrupt parent device for node if there is one
    or otherwise aborts if fatal is true
* Device_lookup::get_or_create_ic(node, ic_ptr)
  * Ic_ok if there is a valid interrupt parent, which is returned via
    *ic_ptr
  * Ic_e_* otherwise

If a valid pointer is returned it may be used to register interrupts.
If the interrupt parent is not a virtual device (Ic_e_no_virtic), then
interrupts are handled by a physical device and can be savely ignored.

Change-Id: I28f11932cb558a9a6d92c5d744a508b4dacfb475
Requires-Change-Id: I8c328abcb24a4fb2de539686c066c2c1d63bb208