Invoke init_device() directly after device creation
Device creation and completion of initialization was seperated. In
a first step all devices were created and their mmio resources were
mapped. In a second step init_devices() was invoked for all devices.

We intend to initialize the devices completely during creation. As
a first step we invoke init_device() in create() and will remove
init_device() on by one in the following commits.

Change-Id: I33d3b4e9d95a8f0c3ef9d748f5e05fe868511914