fix gcc's format-security warnings
'format-security' is a gcc warning option. The error print functions
generate a lot of these warnings because the compiler is unable to check
the provided string and / or cannot find format arguments.

This change fixes these warnings by providing more specialized template
functions which the compiler can better match to the callers. For cases
where no format arguments are provided the implementation directly
prints the provided string.

Change-Id: I579cdc46253a7fafa7da07123058c415ad72fc5c