amd64: redesign Apic_array -> Lapic_array & Io_apic
Up to now, the local APIC was the interrupt parent of the hardware or
virtual devices (PIT). However, the component handling device
interrupts in an SMP system is the IO-APIC and not the local APIC.
While SMP is still unsupported, the new Io_apic handles the distribution
of interrupts to the correct LAPIC.

Another issue is MMIO access to the local APIC as the region is the
same for each vCPU but can only be registered once in the memory map. To
facilitate access to the correct LAPIC in a multi-core setup, the
Lapic_array distributes MMIO read/write accesses to the correct LAPIC.

This basically splits the former Apic_array into two separate component.

Change-Id: I694cab69621b6b12f6bf7a2e04a8ee520093a379