Remove init_device() from ARM Core_timer
The timer was handled in a special way:
* instantiated before device tree traversing
* added to device list iff a matching device node was found
* initialized via init_device()
If there was no matching device node the timer ended up uninitialized
and led to a crash when the uvmm tried to inject a timer interrupt.

Here we restructure the code. We collect irq related information in
create, pass it to the constructor and set the timer reference via
set_timer(). To ensure a valid timer we add an additional check
before starting the guest.

Change-Id: Icf83a3f9cffc048edf3437b377afeea9f7c292a5