Remove init_device() from virtio_{console,proxy)
Here we remove the split of initialization between create()
and init_device by moving code to the constructor and to
create(). We collect all necessary information in create()
and then construct the device.

We leave the initialization of the event connector as a
seperate step instead of initializing it directly while
constructing the device to allow for simple future changes
to the interrupt handling in event connector.

We initialize Device::_device in the constructor to provide
a valid capability to all Device methods.

Change-Id: I7692fd98c0f85d170f13bc1dd2e23cfe4debd958