proper RAM management for population of guest RAM
Manage memory where uvmm wants to add data with a free list to be
more flexible when placing data. At the moment only the first
defined memory region will be available for data, so that it
is possible to control where the data goes to. Ramdisk and
DTB are now moved as far back as possible to leave as large
continuous chunks as possible.

Change-Id: I6e31335856de40ed981560547032a23b62175e49