mmio-proxy: adapt size in reg property to real size
If a dataspace is mapped and the size in the property is larger
than the size of the dataspace, reduce the size advertised in
the reg property.

This allows to give a maximum size in the device tree and then
adapt the actual required size by allocating a dataspace
accordingly.

Change-Id: I9c1c192204697014ac7af19c82a2dd95ff59b139