Mark allocated devices
We find devices on the vbus using different methods:
* lookup based on hid (like GIC, PCI, ICU)
* iteration over all devices on the vbus
To ensure that devices are not handled by different parts of the code we
mark allocated devices by setting their _proxy member (this part got lost
some revisions ago).

To be able to change the _proxy member while keeping the other members
const we make all members private and introduce getter/setter methods
and a constructor.

Change-Id: I7808709bc6bd34b8ca17aa735c3ab0ab5dadf1e5