improve handling of RAM base address
Given that uvmm can handle multiple RAM areas now, there is
no real RAM base address, so get rid of that name. For
binary loading we use the first address of the RAM part available
for loading data into. Make that clear with the naming.

Also change load_to_raw() to use an absolute address for the
same reason. ARM was the only one actually working with offsets
from RAM base, so move address computation there. All other
architectures actually worked with absolute addresses and the
assumption that the RAM base is at 0.

Change-Id: Ia62beefbd26eafacb3801198cc8ff8232859240b