introduce explicit Guest_addr type
This replaces the frequently used L4virtio::Ptr. Guest_addr
is a raw address only and needs Vm_ram::Guest2host() to be
converted into a pointer type. It does have operations for
simple address arithmetics, so that most type-casts are now
unnecessary.

Also changes Region to use the new type and moves Region into
the Vmm namespace.

This change only replaces uses of L4virtio::Ptr and code that
is directly dependend. There are more places where Guest_addr
should be used but these can be changed in subsequent smaller
changes.

The changes revealed some inconsistencies in the use of boot
address and guest address in the amd64 architecture. These
are fixed as well.

Change-Id: I489e09e053881901c32fdb004c2e19a6b5c7493e