amd64: process RH and DM separately
The redirection hint and the destination mode are two independent bits
paresed by two different hardware units. The redirection hint tells a
translation unit - located before the system bus - to do lowest
interrupt priority arbitration between the local APICs to determine the
APIC to send the interrupt to and to rewrite the Destination ID field
accordingly.
The Destination Mode bit is parsed by the local APIC when the message
hits the system bus and the DID field is then interpreted as logical or
physical address.

So the MSI msg first hits the translation unit, which interprets the RH
bit and then places the MSI in on the system bus in the right format.
This system bus message contains the DM bit. The local APIC then parses
this message including the DM bit and acts accordingly.

Change-Id: I5a85891d052cc7e4021227a5b31d51ae20e58667