amd64: allow the initrd above 4G
The xloadflags and ext_ramdisk fields in the linux' boot_params are
used and the boot protocol is set to version 2.12, if these fields are
used.

To test this as memory node must be configured in the device tree and
more than 4GB of memory must be provided in the uvmm-start-script.
Beware, the device tree NEED NOT define memory at 0xfee0'0000 as this is
the APIC-page memory.

fixes ticket #950

Change-Id: If22c1184cbcfdc9e3239c5077025c93f406df13b