generalize IC interface for reading interrupts from DT
The current interface dt_num_interrupts/dt_get_interrupts
works with the assumption that a device has only one
interrupt controller, which is fully able to independently
interpret the interrupt information in the device node.
This no longer holds for the interrupt-extended property.
Interrupt parents may be mixed and consequently the list
of interrupts may have entries of different lengths for
each interrupt.

Change the interface to simply read interrupt information
from a given property list and return the number of fields
read.

Change-Id: Ic1ef2e23d4c86873f2572cdebaddc21278427119