amd64: use bitfield members with APIC timer register
The local APIC timer uses CXX_BITFIELD functions to evaluate and change
the bitfield. It is initialized when the guest writes to the LAPIC's
timer register.

Furthermore, the timer keeps counting when masked and sets the pending
bit, when a masked counter triggers. The interrupt is triggered when the
masked bit is cleared.

Change-Id: Ia5b62d1c93c53109c737312ef4b3a33d14349a73