Introduce MSR device and MSR device map
Devices like the local APIC or the Vmx_state have their own set of MSRs
which are handled by the Virt_lapic and Vmx_state objects.
These devices must register with the VMM to be run checked on an MSR
exit.

To support SMP in the future, the Lapic_array serves as distributor of
an MSR access to the LAPIC of the exiting vCPU. The same purpose serves
the Vcpu_msr_handler.

The MSR device interface and usage is the same as for to the existing
Io_device and Mmio_device.

New interfaces:
  Guest::register_msr_device()
  Msr_device::read_msr()
  Msr_device::write_msr()
  Vm_state::read_msr()
  Vm_state::write_msr()

New types:
  Msr_device
  Vcpu_msr_handler

Replaced interfaces:
  Vm_state_vmx::handle_exec_rmsr() -> read_msr()
  Vm_state_vmx::handle_exec_wmsr() -> write_msr()

Change-Id: I89722de56c11f9772345ac87a446628753326d53