amd64: PIT: compute counter values from tsc
Up to now, the PIT counter decreased when a thread running arbitrarily
send a notification (see timer.h). As the PIT runs with a defined
frequency we can compute the number of PIT-ticks passed using the TSC
difference between starting the counter and now.

This assumes the TSC to be running at constant rate even when power
states are switch, which is not necessarily supported by the CPU, but
this seems to be less bad than before.

Change-Id: I3ef3666b0d2a87ea5579441eba35bd5737a5dcc5