amd64: add KVM clock device and inform guest Linux
The KVM clock is a paravirtualized clock source for a Linux guest.
Linux checks the CPUID.01.ECX[31] - Hypervisor bit to determine if it is
running on a hypervisor. Only then it tries to issue KVM clock CPUID
calls.

The KVM clock determines the CPU speed correctly even when the L4Re host
system already runs virtualized in QEMU. If the L4Re host runs on
physical hardware, the PIT-TSC calibration is correct as well an can
eliminate the need for the KVM clock an hence the need to inform the
guest about the hypervisor.

The guest must be built with the following configuration options:
CONFIG_HYPERVISOR_GUEST=y
CONFIG_KVM_GUEST=y
CONFIG_PARAVIRT_CLOCK=y
CONFIG_HAVE_KVM=y
CONFIG_PTP_1588_CLOCK_KVM is not set

Note: KVM calls besides the KVM clock are unhandled and lead to failure
in the uvmm, e.g. vmcall 0x9 for the PTP_1588_CLOCK_KVM.

Change-Id: If54de477fba7c8a57cc52b9ddaae4f3a2777751b