arm: move GICC installation into an mmio handler
The main advantage of the mmio handler is that the region
will be blocked and uvmm omits an error when another device
is accidentally mapped to the same region. As a side effect,
it also allows to detect when there is a program error and
the GICC page disapears from the task.

Change-Id: I774403066fa63da9ab2b765c00556e9bce1eb56e