ARM: Add support for VPIDR
ARM virtualization allows to define a virtualization processor id which
is returned when the guest reads the MIDR register. Here we add support
which allows the user to set this value by adding a "vpidr" element to the
cpu node in the device tree.
For a description of the possible values see the documentation of MIDR.

Change-Id: I58ebc499e08c46575896549f957d6698826ed68b
Requires-Change-Id: I70d210019c805f989476be8748aefa255d348134