improve error handling for default RAM setup
First, tell the user that there are no memory regions so we are trying
to set up a default RAM region. This requires a "ram" capability. If
that capability isn't passed, don't present a misleading

    VMM: FATAL: Insufficient memory

message but show a proper error message. Also changed the error code to
-L4_ENOENT because that one makes more sense here. While being here,
change the existing Dbg().printf() line to warn.printf() to forcing it
to show the subsys name ("ram"). This is the result with -v enabled:

    VMM[ram]: No memory nodes found, setting up default region
    VMM: FATAL: Grabbing default "ram" capability: No such object found

which looks better.

Change-Id: I9cb90a4200af5b8ba398328b0ca535bca5107c05