ARM: Add support for suspend related PSCI calls
To suspend the system guests issue the following PSCI calls:
* cpu_off to switch off aditional CPUs
* psci_affinity_info to check whether all CPUs are off
* system_suspend

Here we add the missing support for cpu_off and psci_affinit_info
and adjust the other code to support these operations.

CPU_OFF

To switch of CPUs we put the underlying thread to sleep. To re-activate
it again it must be forced to continue at some other place.
To be able to check the state of a CPU we add an online_state
member to Cpu_dev and update it when switching CPUs on and off.

AFFINITY_INFO

With PSCI v1.0 guests query single CPUs. So we simply lookup the CPU
based on the provided hardware id and return the online state.

SYSTEM_SUSPEND

To be able to actually suspend the system we have to make sure all
other virtual CPUs are offline before suspending the system. We
restrict the suspend operation to CPU0.

CPU_ON

The methods activation virtual CPUs are now invoked in to different
situations:
* for a newly created CPU
* for a suspended CPU

For a suspended CPU we have to get the underlying thread out of its
sleep loop. We add a small piece of trampoline code used to force
the thread into executing Cpu_dev::reset().

Change-Id: Ic759cf8d774caf6a1cbe9f5b40a1fee2fde11de0