ARM/PSCI: fix PSCI feature function for unknown funcs
The feature function should not just look at the last two bytes when
comparing the function id. A function id can be 32 bits long and
therefore we need to make sure to compare all bits (see "5.1.14
PSCI_FEATURE" in the PSCI platform design document (Issue D)).

A function id is formed by the smc calling convention and we therefore
have to make sure to take everything into account. See the ARM SMC
calling convention document for the bit definitions.

This problem was noticed because when a function id of 0x8000000 was
used, the function would return success. The reason being 0x8000000 &
0xff equals zero which we support (Psci_version). However, 0x8000000 is
a completely different function (actually it is in the ARM Architecture
Calls range)

See https://lkml.org/lkml/2018/1/29/448 which introduced a linux change
showing this wrong behaviour on our side.

Change-Id: I13159301e5110cef24caf96a59df7e8dc0c38657