ARM: Fix validity check for GICC page
Some device trees specify regions smaller then one page for the
GICC region. Therefore we have to relax the current check which
enforces a size of at least one page.

We now check whether the page provided by the kernel is able to
back the region specified in the device tree. That means that
the region specified in the device tree must fit into one page.

TODO:
The actual GICC page has a member at offset 0x1000 - GICC_DIR.
To support access to this field the kernel would have to map
two pages and uvmm would have to accept regions with a size of
up to 0x2000. Since our primary guest OS Linux does not use
this field we currently do not support access to GICC_DIR. If
that changes we have to adjust this test again.

Change-Id: Ie754ecbc98801e68235442fae2bee139b88333ed