ARM: Adapt GICC reg entry size and remove superflous ones
We currently only support GICC region size up to 0x1000. There
are device trees which specify either
* 0x2000 to provide access to GICC_DIR
* 0x20000 to support GIC versions which provide non banked GICC
  registers and therefore need some special handling

When we leave a size larger than 0x1000 in the device tree the guest
might try to access the complete region which would lead to
termination of uvmm. To be able to support upstream device trees
we adjust the size of the GICC reg entry and remove any VGIC related
entries.

Change-Id: I3af350c5df76813252e584534468b1ac14ab469a