Fix iteration over reg entries
In Change-Id: I413b0d004d14209c4e3674d29c8b0ab60676f4f2 we
re-structured the code iterating over the reg entry and directly
returned the result of mmio_region_valid() from within a loop.
This broke the iteration over all reg entries.

Here we re-establish the previous behaviour and only return
when mmio_region_valid() signals an error or when we reach
the end of the reg entries.

Change-Id: Ia8bb8757e88bf95b0cb9bf16ea48cbe99f4c73b4