Remove mmio fallback
Since Change-Id:Ice0eb0669fe994f38fb0a93d87be9f9cca6ba3c5 we add
every resource available on the vbus to the address space of the
VM. If the VM raises a valid pagefault there should be a valid
region in the memory map. If there is none there is also no entry
on the vbus providing the missing resource. Therefore we can
remove any attempt to resolve the pagefault by asking io about the
missing resource.

Here we remove everything related to the mmio fallback.

Change-Id: I7003d8d6ecdad8445e09d40e3e65b990849ea7c8