Move smc/hvc handler into its own devices
The handling of smc/hvc calls is unified and more than one handler is
possible for both call types.

Implement a PSCI device which is only created if the corresponding dt
entry is available. This supports now also smc for SMCCCC in addition to
the hvc call. The new device can be disabled at compile time.

The vm print functionality moved into its own device which is always
available.

All PSCI tests are changed to add the now necessary dt entry.

Change-Id: I702fd110fbc9baa026d519795cdf7a0b1ae13212