io_proxy: correctly handle ret=1 from IRQ bind
According to the documentation of L4::Icu::bind(), a return value of 0
means a direct unmask via the IRQ object using L4::Irq::unmask(). This
is the normal case and currently implemented.

But a return value of 1 means that the interrupt has to be unmasked via
the ICU using L4::Icu::unmask().

Change-Id: I2bd8d94a0ecb1c905718594eef73158ef295fb66