virt_bus: support MMIO space resources
In certain cases an MMIO resource cannot be mapped to the client, for
instance, if the MMIO resource provider needs to trap-and-emulate. Uvmm
provides MMIO resources to guests this way.

The IO server in turn can provide MMIO space resources to its clients.
A client can distinguish between a normal MMIO resource and an MMIO
resource with an MMIO space handler by checking the resource flags. For
MMIO spaces, any read/write access is translated into the corresponding
mmio_read() / mmio_write() operation of the Mmio_space handler.

Change-Id: I7ade8d20243bb413156faeef61aa0183833b8422
Requires-Change-Id: I24f47433a1cac2e76bff7014ea2a7100a5e23b03