Virq: Add diagnostic message to failed register_obj()
If the initial environment is not setup correctly the cap slot
referenced by l4vmm,virqcap might point to an invalid cap (e.g.
a void cap). In this case register_obj might fail and we and up
with a simple Fatal: Insufficiant memory.

Here we add a better diagnostic message for this case.

Change-Id: Ibc38505ca248fccdedae8806919aff437e9d3b37