ARM: Protect against invalid affinity values in device tree property
The reg entry of a CPU node specifies the affinity of a CPU. The value
is used to set the VMPIDR register and to address CPUs in PSCI calls.

To ensure that we use correct values in both contexts we apply the
affinity mask directly after reading the reg property.

Change-Id: I2ab68e2609f1dd884ae26d8864c1630177ab0693