ARM/PSCI: Handle 32bit affinity arguments
The 32bit ABI places Aff3 in bit [31-24], the MPIDR register places it
int bit [39-32]. If we see a 32bit PSCI call on a 64bit system we have
to convert the parameter and move aff3 from bit [31-24] to bit [39-32]
in PSCI call with an affinity as argument.

Here we introduces an affinity() method handling the conversion and
adapt Affinity_info and Cpu_on to apply this method on its arguments.

Change-Id: I1d4fbb91f22622c980aaaeb70642762847a7e373