add skeleton for virtio-rng device
The device forwards the random number request to an
appropriate L4 random number generation service.
At the moment the device itself is not implemented
yet, only setting of a kaslr seed value is available.

As the device needs to modify other nodes than its own
in the device tree, a post-processing hook for devices
is introduced.

Change-Id: Ic024a7fbaad9e52b3efbac43fab0b7cc7b1291ac
Requires-Change-Id: I38e516aad74fb113acfe3c7e3f9be42b31638451