Enhance `cpu` monitor command
This creates `server/src/monitor/ARCH-*` directories which contain architecture
specific monitoring code (in particular register dumping functions) that was
previously located under `server/src/ARCH-*`.

The `cpu` monitor command is also extended such that it is now possible to
list all available cpus and dump either the registers of one specific cpu
or of all of them at once (prior to this change only the latter was possible).

Change-Id: I360f677b449e5ca9d4c3021c2c4365f3f8b99e70