amd64: fix Application Processor initialization
In Generic_cpu_dev::powerup_cpu(), vCPUs with an id > 0 are set up to
run Guest::startup() on a new thread but not started yet. The
Guest::run() method was calling Generic_cpu_dev::startup()
on the main thread unconditionally for any vCPU, causing the VM to enter
a fatal state.

Change-Id: Ida0f57f48a3b890bc98e61a31a7a1b00c76e42eb